+++ src/atlas/zimgautothreshold.cpp
@@ -146,7 +146,7 @@ TVoxel ZImgAutoThreshold<ReportProgress>::typedCentroidThre(double &cent1, doubl
     prevThreBin = threBin;
 
     double totalPos = std::accumulate(weightPos.begin()+low, weightPos.begin()+threBin+1, .0);
-    size_t totalWeight = std::accumulate(hist.begin()+low, hist.begin()+threBin+1, size_t(0));
+    size_t totalWeight = std::accumulate(hist.begin()+low, hist.begin()+threBin+1, 0_usize);
     if (totalWeight == 0) {
       cent1 = (threBin + low) / 2.0;
     } else {
@@ -154,7 +154,7 @@ TVoxel ZImgAutoThreshold<ReportProgress>::typedCentroidThre(double &cent1, doubl
     }
 
     totalPos = std::accumulate(weightPos.begin()+threBin+1, weightPos.begin()+high+1, .0);
-    totalWeight = std::accumulate(hist.begin()+threBin+1, hist.begin()+high+1, size_t(0));
+    totalWeight = std::accumulate(hist.begin()+threBin+1, hist.begin()+high+1, 0_usize);
     if (totalWeight == 0) {
       cent2 = (high + threBin) / 2.0;
     } else {
