+++ src/img/zimgometiff.cpp
@@ -94,7 +94,7 @@ void ZImgOmeTiff::writeImg(const QString &filename, const ZImg &img, Compression
 {
   ZTiffWriter tiffWriter;
   int extraSample = img.info().lastChannelIsAlphaChannel ? 2 : -1;  //EXTRASAMPLE_UNASSALPHA or none
-  if (img.byteNumber() > size_t(1024)*1024*3600)
+  if (img.byteNumber() > 1024_usize*1024*3600)
     tiffWriter.startWriting(filename, comp, extraSample, true);
   else
     tiffWriter.startWriting(filename, comp, extraSample, false);
@@ -117,7 +117,7 @@ void ZImgOmeTiff::writeImg(const QString &filename, const ZImgSliceProvider &img
 {
   ZTiffWriter tiffWriter;
   int extraSample = imgSliceProvider.imgInfo().lastChannelIsAlphaChannel ? 2 : -1;  //EXTRASAMPLE_UNASSALPHA or none
-  if (imgSliceProvider.imgInfo().byteNumber() > size_t(1024)*1024*3600)
+  if (imgSliceProvider.imgInfo().byteNumber() > 1024_usize*1024*3600)
     tiffWriter.startWriting(filename, comp, extraSample, true);
   else
     tiffWriter.startWriting(filename, comp, extraSample, false);
